Update RulerValuePicker.java
mHorizontalScrollView isn't fully initialized at first to accept smoothScrollView.
So there was a 400ms delay for each selectValue(). 

The fix i made will selectValue() instantly without annoying delay and in cases when it fails to do that - it will loop 3 attemps with 100ms delay each trying to set (which should be enougth time for the mHorizontalScrollView to initialize (kinda a fallback to old functionality but with attempts to scrool as soon as possible).